+++ src/i2c/mod.rs
@@ -4,9 +4,21 @@ use std::sync::mpsc::Sender;
 
 //use serde::{Serialize, Deserialize};
 
-use crate::Event;
+use crate::event::Event;
 
 use crate::protocol::*;
+
+
+#[derive(Clone, Debug)]
+pub enum I2CEvent {
+    On(Button),
+    Off(Button),
+}
+
+
+pub struct I2CStruct {
+    pub tx: Sender<Event>
+}
 
 pub fn initialize(tx: Sender<Event>) -> I2CStruct {
     I2CStruct {
